+++ CHANGELOG.md
@@ -60,6 +60,8 @@
 
 - Add serialization of modules and services on app initialization (INDIGO Sprint 210903, [!170](https://github.com/TeskaLabs/asab-webui/pull/170))
 
+- Add reusable `KnowledgeBase` component (INDIGO Sprint 210917, [!176](https://github.com/TeskaLabs/asab-webui/pull/176))
+
 ### Refactoring
 
 - Renaming configuration option FAKE_USERINFO to MOCK_USERINFO and refactoring code accordingly (INDIGO Sprint 210406, [!91](https://github.com/TeskaLabs/asab-webui/pull/91))